Correctly complete parsing when no further input
The `complete` method handles the case when no further input will be
available by converting `Continue` to `Success`. Large changes through
the code base to parameterize control in `loop` to either `complete`, or
`parse` as normal.

Updated to Scala 3.0.1 to fix some compiler bugs. Had to turn off
`sbt-tpolecat` to get past an erroneous warning on the pattern match in
`loop`.